Move ozw migration tests to own module
